requirements.txt
setuptools==36.2.0
pytest==3.2.1
mypy==0.521
pylint==1.7.2
gunicorn==19.7.1
flask==0.12.2
google-api-python-client==1.6.3
httplib2==0.10.3
oauth2client==3.0.0
google-cloud==0.27.0
pytest-flask==0.10.0
